requirements.txt
alembic==1.13.1
anyio==3.7.1
asyncpg==0.27.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
databases==0.6.2
distlib==0.3.8
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
nodeenv==1.8.0
ormar==0.12.2
packaging==24.0
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.8
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.41
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.10.0
uvicorn==0.29.0
virtualenv==20.25.1
PyJWT